Move ToJS/FromJS classes to inline-js-core
