+++ inline-js-core/src/Language/JavaScript/Inline/Core.hs
@@ -8,21 +8,16 @@ module Language.JavaScript.Inline.Core
     newSession,
     closeSession,
 
-    -- * Evaluation requests
+    -- * Haskell/JavaScript data marshaling
     JSExpr,
     JSVal,
-    code,
-    buffer,
-    string,
-    json,
-    jsval,
+    EncodedJSON (..),
+    ToJS (..),
+    FromJS (..),
 
     -- * Performing evaluation
     -- $notes-eval
-    evalNone,
-    evalBuffer,
-    evalJSON,
-    evalJSVal,
+    eval,
     importCJS,
     importMJS,
 
@@ -33,10 +28,8 @@ module Language.JavaScript.Inline.Core
   )
 where
 
-import qualified Data.ByteString.Lazy as LBS
-import Data.String
+import Language.JavaScript.Inline.Core.Class
 import Language.JavaScript.Inline.Core.Exception
-import Language.JavaScript.Inline.Core.Instruction
 import Language.JavaScript.Inline.Core.JSVal
 import Language.JavaScript.Inline.Core.Message hiding
   ( code,
@@ -44,27 +37,9 @@ import Language.JavaScript.Inline.Core.Message hiding
 import Language.JavaScript.Inline.Core.Session
 import System.Directory
 
--- | Convert a 'String' to 'JSExpr'. In most cases, using the 'IsString'
--- instance with the @OverloadedStrings@ extension is more convenient.
-code :: String -> JSExpr
-code = fromString
-
--- | Embed a 'LBS.ByteString' as a @Buffer@ expression.
-buffer :: LBS.ByteString -> JSExpr
-buffer = JSExpr . pure . BufferLiteral
-
 -- | Embed a 'String' as a @string@ expression.
 string :: String -> JSExpr
 string = JSExpr . pure . StringLiteral
-
--- | Embed a UTF-8 encoded JSON expression. It will be parsed with
--- @JSON.parse()@.
-json :: LBS.ByteString -> JSExpr
-json = JSExpr . pure . JSONLiteral
-
--- | Embed a 'JSVal' as an expression.
-jsval :: JSVal -> JSExpr
-jsval = JSExpr . pure . JSValLiteral
 
 -- $notes-eval
 --
@@ -96,7 +71,7 @@ jsval = JSExpr . pure . JSValLiteral
 importCJS :: Session -> FilePath -> IO JSVal
 importCJS s p = do
   p' <- makeAbsolute p
-  evalJSVal s $ "require(" <> string p' <> ")"
+  eval s $ "require(" <> string p' <> ")"
 
 -- | Import an ECMAScript module file and return its module namespace object.
 -- The module file path can be absolute, or relative to the current Haskell
@@ -104,7 +79,7 @@ importCJS s p = do
 importMJS :: Session -> FilePath -> IO JSVal
 importMJS s p = do
   p' <- makeAbsolute p
-  evalJSVal s $
+  eval s $
     "import('url').then(url => import(url.pathToFileURL("
       <> string p'
       <> ")))"
