+++ inline-js/src/Language/JavaScript/Inline.hs
@@ -2,14 +2,8 @@ module Language.JavaScript.Inline
   ( -- * Core functionalities
     module Language.JavaScript.Inline.Core,
 
-    -- * Haskell/JavaScript data marshaling type classes
-    ToJS (..),
-    FromJS (..),
+    -- * @aeson@ support
     Aeson (..),
-    EncodedJSON (..),
-
-    -- * Polymorphic eval function
-    eval,
 
     -- * QuasiQuoters for inline JavaScript
     expr,
@@ -17,6 +11,6 @@ module Language.JavaScript.Inline
   )
 where
 
-import Language.JavaScript.Inline.Class
+import Language.JavaScript.Inline.Aeson
 import Language.JavaScript.Inline.Core
 import Language.JavaScript.Inline.TH
